remove deployment of custom powershell script for this role
The feature to always run at deployment time and run once only is in role
win_base with inventory variable custom_powwershell_commands. Thus we will move
the script here into inventory and remove the deploy/run it.

It is better to keep the role focus on iis and move things that should belong
to inventory to, well inventory.