allow-to-skip-install-base
This flag allows us to install the win_iis_base_features and releated
chocolatey packages. This usually only requires when we build the instance
AMI or setup the server the first time - or upgrade the packages to the new
versions etc..

By setting this to false the role will skip it and focusing on setting up and
configure the websites and application pool which save a lot of time.