Re-implementation of win_iis websites and apppool
The current ansible modules has some limitations which due to the inconsistent
behaviour of powershell for iis. The issues are:
- The apppool created does not have the correct Identity
- The sites created always having http binding and does not get cleanup properly

This change will use a simple powershell to get the job done and not tweaking
ansible module. The goal is to make it simple thus if it does not work with
powershell update/etc it is much easier to update it and test.

The current implementation of ansible modules seems to be very over-engineered.