Use powershell to install win features
The ansible module can not install it, always return "Failed to install Windows
Feature: ArgumentNotValid: The role, role service, or feature name is not
valid"

but I can install it using that powershell snipet.

Seems ansible for windows is not well maintained in terms of bug fix - probably
just stick with powershell now and these powershell command are idempotant
anyway.